field/index loads and stores

+++ ex/vm3/src/ast.rs
@@ -188,7 +188,7 @@ pub struct SetVar<'arena, 'src> {
 
 #[derive(Clone, Debug)]
 pub enum Key<'src> {
-  Int(usize),
+  Int(i32),
   Ident(Ident<'src>),
 }
